Introducing new `Checkout::Update`
Now automatically sets the proper cart step so frontend devs don't need to worry about it anymore.
